new helper module
fixes: https://github.com/theforeman/foreman-ansible-modules/issues/47
