+++ module_utils/foreman_helper.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+# (c) Matthias Dellweg (ATIX AG) 2017
+
+import re
+import yaml
+
+
+def handle_no_nailgun(module, has_nailgun):
+    if not has_nailgun:
+        module.fail_json(msg="Missing required nailgun module (check docs or install with: pip install nailgun)")
+
+
+# Helper for templates
+def parse_template(template_content, module):
+    try:
+        data = re.match(
+            '.*\s*<%#([^%]*([^%]*%*[^>%])*%*)%>', template_content)
+        if data:
+            datalist = data.group(1)
+            if datalist[-1] == '-':
+                datalist = datalist[:-1]
+            template_dict = yaml.safe_load(datalist)
+        # No metadata, import template anyway
+        template_dict['template'] = template_content
+    except Exception as e:
+        module.fail_json(msg='Error while parsing template: ' + str(e))
+    return template_dict
+
+
+def parse_template_from_file(file_name, module):
+    try:
+        with open(file_name) as input_file:
+            template_content = input_file.read()
+            template_dict = parse_template(template_content, module)
+    except Exception as e:
+        module.fail_json(msg='Error while reading template file: ' + str(e))
+    return template_dict
