+++ modules/katello_upload.py
@@ -84,6 +84,10 @@ try:
 except:
     HAS_NAILGUN_PACKAGE = False
 
+from ansible.module_utils.basic import AnsibleModule
+from ansible.module_utils._text import to_native
+from ansible.module_utils.foreman_helper import handle_no_nailgun
+
 
 class NailGun(object):
 
@@ -151,8 +155,7 @@ def main():
         supports_check_mode=True,
     )
 
-    if not HAS_NAILGUN_PACKAGE:
-        module.fail_json(msg="Missing required nailgun module (check docs or install with: pip install nailgun")
+    handle_no_nailgun(module, HAS_NAILGUN_PACKAGE)
 
     server_url = module.params['server_url']
     username = module.params['username']
@@ -185,9 +188,6 @@ def main():
 
     module.exit_json(changed=True, result="File successfully uploaded to %s" % repository)
 
-
-from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils._text import to_native
 
 if __name__ == '__main__':
     main()
